resources refactoring
