+++ src/app/App.tsx
@@ -5,9 +5,8 @@ import Routes from 'app/Routes';
 import NormalizedStyles from 'components/styles/NormalizedStyles';
 import BaseStyles from 'components/styles/BaseStyles';
 import { store } from 'app/store';
-import { theme } from 'resources/theme';
+import { theme } from 'app/theme';
 
-import 'resources/css/fonts.css';
 import 'rc-tooltip/assets/bootstrap.css';
 
 function App() {
