--- src/resources/css/fonts.css
@@ -1,7 +1,0 @@
-@font-face {
-  font-family: 'jira';
-  src: url('../fonts/jira.woff') format('truetype'), url('../fonts/jira.ttf') format('woff'),
-    url('../fonts/jira.svg#jira') format('svg');
-  font-weight: normal;
-  font-style: normal;
-}
