--- src/resources/theme/colors.ts
@@ -1,20 +1,0 @@
-export const colors = {
-  primary: '#0052cc',
-  success: '#0B875B',
-  danger: '#E13C3C',
-  warning: '#F89C1C',
-  secondary: '#F4F5F7',
-  textDarkest: '#172b4d',
-  textDark: '#42526E',
-  textMedium: '#5E6C84',
-  textLight: '#8993a4',
-  white: '#FFFFFF',
-  //
-  borderLightest: '#dfe1e6',
-  borderLight: '#C1C7D0',
-  borderInputFocus: '#4c9aff',
-  backgroundDarkPrimary: '#0747A6',
-  backgroundLight: '#ebecf0',
-  backgroundLightPrimary: '#005eeb',
-  backgroundLightSuccess: '#E4FCEF',
-};
